refactor: change package name
Change "com.bignerdranch.android.tbnr_geoquiz" to "com.bignerdranch.android.geoquiz"
