+++ app/src/androidTest/java/com/bignerdranch/android/geoquiz/ExampleInstrumentedTest.kt
@@ -1,12 +1,10 @@
-package com.bignerdranch.android.tbnr_geoquiz
+package com.bignerdranch.android.geoquiz
 
-import androidx.test.platform.app.InstrumentationRegistry
 import androidx.test.ext.junit.runners.AndroidJUnit4
-
+import androidx.test.platform.app.InstrumentationRegistry
+import org.junit.Assert.*
 import org.junit.Test
 import org.junit.runner.RunWith
-
-import org.junit.Assert.*
 
 /**
  * Instrumented test, which will execute on an Android device.
@@ -19,6 +17,6 @@ class ExampleInstrumentedTest {
     fun useAppContext() {
         // Context of the app under test.
         val appContext = InstrumentationRegistry.getInstrumentation().targetContext
-        assertEquals("com.bignerdranch.android.tbnr_geoquiz", appContext.packageName)
+        assertEquals("com.bignerdranch.android.geoquiz", appContext.packageName)
     }
 }
