+++ app/src/main/java/com/bignerdranch/android/geoquiz/Question.kt
@@ -1,4 +1,4 @@
-package com.bignerdranch.android.tbnr_geoquiz
+package com.bignerdranch.android.geoquiz
 
 import androidx.annotation.StringRes
