+++ app/src/main/java/com/bignerdranch/android/geoquiz/QuizViewModel.kt
@@ -1,4 +1,4 @@
-package com.bignerdranch.android.tbnr_geoquiz
+package com.bignerdranch.android.geoquiz
 
 import androidx.lifecycle.SavedStateHandle
 import androidx.lifecycle.ViewModel
