+++ app/src/test/java/com/bignerdranch/android/geoquiz/ExampleUnitTest.kt
@@ -1,8 +1,7 @@
-package com.bignerdranch.android.tbnr_geoquiz
+package com.bignerdranch.android.geoquiz
 
+import org.junit.Assert.assertEquals
 import org.junit.Test
-
-import org.junit.Assert.*
 
 /**
  * Example local unit test, which will execute on the development machine (host).
